Digimon: Patamon Sequence v0.1.1
Modified selection of In-Training Tokomons to fill the Digi-egg deck.

Levelled out the ratios of each level of Digimon to make the early-game easier. Previously, my hand would be filled with level-6s, but no level-5s would be drawn. ArkhaiAngemon was chosen as few other level-5s matched well with the angelic deck; the choice will likely be revised later.

Removed unnecessary cards, those being Cherubimon (only usefull when multiple Digimon are suspended) and Ophanimon (it just returns to the security stack on deletion, which serves little use in progressing forward without a card to delete it).